requirements.txt
anyio==3.3.4
asgiref==3.4.1
asyncpg==0.24.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
boto3==1.20.12
botocore==1.23.12
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cycler==0.11.0
databases==0.5.3
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.0
fonttools==4.28.2
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.1.0
kiwisolver==1.3.2
lxml==4.6.4
MarkupSafe==2.0.1
matplotlib==3.5.0
numpy==1.21.4
orjson==3.6.4
packaging==21.3
pandas==1.3.4
passlib==1.7.4
Pillow==8.4.0
psycopg2==2.9.1
pycparser==2.20
pydantic==1.8.2
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.1
python-multipart==0.0.5
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
s3transfer==0.5.0
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.3
SQLAlchemy==1.4.26
starlette==0.16.0
threadpoolctl==3.0.0
tomli==1.2.2
typing-extensions==3.10.0.2
ujson==4.2.0
urllib3==1.26.7
uvicorn==0.15.0
watchgod==0.7
websockets==10.0
xgboost==1.5.1
